added small animations, songs are back here, moved icons to another folder, preloaded all sfx, added contributers
