+++ boards.css
@@ -174,6 +174,9 @@
     left: 110%;
     width: max-content;
     max-width: 40vw;
+    opacity: 0;
+    pointer-events: none;
+    transition: all 0.3s ease;
 }
 
 .redochoice {
