+++ info.md
@@ -5,13 +5,16 @@ TETI (the name is in the works) is a modern tetris clone I've been working on fo
 
 This project started out as a learning experience, because suprisingly it was my first time actually using javascript. For my year 12 major project (pseudohuman) I wanted to build a multiplayer web based game, and thus I made this to gain experience before starting. I had already programmed in Java, as wlel as some C# and python, so I was not unfamiliar to programming, and I already knew frontend basics. Teti was able to combine my interest for tetris, and also challenge me to learn javascript. Well after the project was over, I could truly improve the code and game to make an amazing experience
 
-I hope you enjoy TETI, and give a star if you like it. If you want you can also contribute, and open issues, and I'll try to keep up with new features.
-
 [Here is the trailer video](https://www.youtube.com/watch?v=Gf2dsPRf2uM)
 
-I know in this new update the code is no longer 1000 lines, but I think it's time to let go of this limitation in order to grow and make this a proper tetris client. However, I'll try to keep the spirit of using no dependacies and pure html, css, and javascript.
+I hope you enjoy TETI, and give a star if you like it. The code is no longer 1000 lines, but I want this to grow to a proper maintainable tetris client.
 
-Also can someone help me figure out of the movement feels off or is it just me
+Also is it just me or is the movement a bit off
+
+### contributions
+Thank you [existancepy](https://github.com/existancepy) for all the fixes.
+
+Feel free to contribute with pull requests and issues. I'll try to keep up with new features and fixes.
 
 ## Current Features
 - modern guideline (customisable)
@@ -22,8 +25,8 @@ Also can someone help me figure out of the movement feels off or is it just me
     - srs+ (tetrio)
     - blockout and lockout
     - ghost
-    - tspin
-    - proper locking (with indicator)
+    - tspins and allspins
+    - locking with indicator
 - stats sidebar
 - customisable game settings
 - customisable controls, handling
@@ -41,32 +44,30 @@ Also can someone help me figure out of the movement feels off or is it just me
 - menu animations
 - import/export settings
 - customise queues
-- allspin (immobile check like tetrio)
 - edit and draw on board (custom game)
-
+- undo and redo with non linear branches
 
 ## Updates
+### v1.1.0
+- added undo and redo feautures
+- non linear timeline: branches can be made and navigated, much like a git tree
+- preloaded all sfx and songs
+
 ### v1.0.0
-I should probably start numbering these versions huh.
-
-Reorganised all the code into very managable classes that can now be more easily changed.
-
-You can now draw on your board in custom game.
-
-You can open an editing menu in custom games (default key: e). This allows you to import, export boards, and also add/remove lines and much more.
-
-The main menu is more organised, and each menu now organises a certain setting group for better code efficiency.
-
-The settings object is changed to be a hashmap (js object) rather than an array, more closely matches tetrio setting.ttc files.
-
+- I should probably start numbering these versions huh.
+- Reorganised all the code into very managable classes that can now be more easily changed.
+- You can now draw on your board in custom game.
+- You can open an editing menu in custom games (default key: e). This allows you to import, export boards, and also add/remove lines and much more.
+- The main menu is more organised, and each menu now organises a certain setting group for better code efficiency.
+- The settings object is changed to be a hashmap (js object) rather than an array, more closely matches tetrio setting.ttc files.
 
 ## TODO list
 Things that I am working on, and most likely will release in the next update
-- preload all audio 
+- allow for modifiers to be added to keybinds
+- loading page for the first second or so
 
 ## Feature Wishlist
 No particular order or time frame, may or may not add these
-- undo/redo
 - gamemode presets
 - guide like progression thing? using custom boards (kinda like tetris tres bien)
 - title screen maybe
@@ -84,3 +85,4 @@ No particular order or time frame, may or may not add these
 - more rotation systems (ars, trs, srs/srsX, none)
 - misdrop remover mode
 - add different 4w residue (both 3 and 6 residue)
+- show info when hovering over settings
